# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceResponse": {
        "description": "A registered device token.",
        "properties": {
          "device_os": {
            "title": "Device Os",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "device_os"
        ],
        "title": "DeviceResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterDeviceRequest": {
        "additionalProperties": false,
        "description": "Register (or refresh) a push token for the caller's device.",
        "properties": {
          "device_os": {
            "description": "IOS or ANDROID.",
            "title": "Device Os",
            "type": "string"
          },
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "device_os"
        ],
        "title": "RegisterDeviceRequest",
        "type": "object"
      },
...
      "UnregisterDeviceRequest": {
        "additionalProperties": false,
        "description": "Remove a push token.",
        "properties": {
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "UnregisterDeviceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Giftly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/devices": {
      "delete": {
        "description": "Remove a push token (only if it belongs to the caller).",
        "operationId": "unregister_device_api_devices_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unregister Device",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "description": "Register or refresh a push token for the caller.",
        "operationId": "register_device_api_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Device",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}